# openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "SmartScheduler API",
    "description": "SmartScheduler API for job scheduling and contractor management",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5004"
    }
  ],
  "paths": {
...
    "/api/recommendations": {
      "post": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Get contractor recommendations",
        "description": "Get ranked contractor recommendations for a job with up to 3 suggested time slots per contractor.",
        "operationId": "GetRecommendations",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/api/recommendations/recalculate": {
      "post": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Recalculate recommendations for a job",
        "description": "Recalculates contractor recommendations for a job, updates the job's recommendation reference, and publishes a real-time update event.",
        "operationId": "RecalculateRecommendations",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecalculateRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RecommendationResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "requestId": {
            "type": "string",
            "format": "guid"
          },
          "jobId": {
            "type": "string",
            "format": "guid"
          },
          "recommendations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecommendationDto"
            }
          },
          "bestRecommendationContractorId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "configVersion": {
            "type": "integer",
            "format": "int32"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RecommendationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "contractorId": {
            "type": "string",
            "format": "guid"
          },
          "contractorName": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "scoreBreakdown": {
            "$ref": "#/components/schemas/ScoreBreakdown"
          },
          "rationale": {
            "type": "string"
          },
          "suggestedSlots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeSlotDto"
            }
          },
          "distance": {
            "type": "number",
            "format": "double"
          },
          "eta": {
            "type": "integer",
            "format": "int32"
          },
          "contractorBaseLocation": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ScoreBreakdown": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "availability": {
            "type": "number",
            "format": "double"
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "distance": {
            "type": "number",
            "format": "double"
          },
          "rotation": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
      "TimeSlotDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "startUtc": {
            "type": "string",
            "format": "date-time"
          },
          "endUtc": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "type": "string"
          },
          "confidence": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "RecommendationRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "jobId": {
            "type": "string",
            "format": "guid"
          },
          "desiredDate": {
            "type": "string",
            "format": "date"
          },
          "serviceWindow": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TimeWindowDto2"
              }
            ]
          },
          "maxResults": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TimeWindowDto2": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RecalculateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "jobId": {
            "type": "string",
            "format": "guid"
          }
        }
      },
...
    }
  }
}